Added player code & entity management
I added some simple and rough code for a player to move, as well as the base for a slightly more complex system for managing "Entities" in the scene/code. I also made a "Tracker" in the code, to fetch Entities.

Set up a somewhat isometric camera, and a textured floor